+++ #633 - Server/combat/com/rs/game/npc/godwars/armadyl/KreeArra.java
@@ -61,7 +61,7 @@ public class KreeArra extends NPC {
 			public void run() {
 				setFinished(false);
 				World.addNPC(npc);
-				npc.setLastRegionId(0);
+				npc.setLastRegionId((short) 0);
 				World.updateEntityRegion(npc);
 				loadMapRegions();
 				checkMultiArea();
