+++ #633 - Server/combat/com/rs/game/npc/godwars/zammorak/KrilTstsaroth.java
@@ -62,7 +62,7 @@ public class KrilTstsaroth extends NPC {
 			public void run() {
 				setFinished(false);
 				World.addNPC(npc);
-				npc.setLastRegionId(0);
+				npc.setLastRegionId((short) 0);
 				World.updateEntityRegion(npc);
 				loadMapRegions();
 				checkMultiArea();
