+++ #633 - Server/src/com/rs/game/player/actions/HomeTeleport.java
@@ -51,7 +51,7 @@ public class HomeTeleport extends Action {
 	    if (player.getControlerManager().getControler() == null)
 		Magic.teleControlersCheck(player, tile);
 	    player.setNextFaceWorldTile(new WorldTile(tile.getX(), tile.getY(), tile.getPlane()));
-	    player.setDirection(6);
+	    player.setDirection((byte) 6);
 	    player.lock(11);
 	} else if (currentTime == 19) {
 	    player.setNextGraphics(new Graphics(HOME_GRAPHIC + 1));
